Feature gate cluster in tests
